* Impot config.jelly definitions. * Move ContainerInfo
