+++ src/main/resources/org/jenkinsci/plugins/mesos/MesosAgentSpecTemplate/ContainerInfo/config.jelly
@@ -1,0 +1,78 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
+    <f:entry title="${%Container Type}" field="type">
+      <f:entry title="${%Docker Image}" field="dockerImage" description="Image to be run under UCR or DOCKER.">
+          <f:textbox clazz="required"/>
+      </f:entry>
+        <f:radioBlock name="type" title="${%Docker}" value="DOCKER" inline="true" checked="${instance.type == null || instance.type == 'DOCKER'}">
+            <f:entry title="${%Docker Privileged Mode}" field="dockerPrivilegedMode" description="This will start the image using Docker's privileged mode.">
+                <f:checkbox/>
+            </f:entry>
+            <f:entry title="${%Docker Force Pull Image}" field="dockerForcePullImage" description="This will force a pull of the Docker Image regardless of whether it exists locally.">
+                <f:checkbox/>
+            </f:entry>
+            <f:entry title="${%Docker Image Can Be Customized}" field="dockerImageCustomizable" description="This will allow override default docker image using labels. E.g.: mesosSlaveLabel:evarga/jenkins-slave:latest">
+                <f:checkbox/>
+            </f:entry>
+        </f:radioBlock>
+        <f:radioBlock name="type" title="${%Mesos}" value="MESOS" inline="true" checked="${instance.type == 'MESOS'}"></f:radioBlock>
+    </f:entry>
+
+    <f:entry title="${%Docker in Docker image}" field="isDind" >
+        <f:checkbox/>
+    </f:entry>
+
+    <f:entry title="${%Use custom docker command shell}" field="useCustomDockerCommandShell" >
+        <f:checkbox/>
+    </f:entry>
+
+    <f:entry title="${%Custom docker command shell}" field="customDockerCommandShell" >
+        <f:textbox field="customDockerCommandShell"/>
+    </f:entry>
+
+    <f:entry title="${%Networking}" description="${%Specify the networking mode to use for this container}" field="networking" >
+        <f:radioBlock id="networking.HOST" name="networking" title="${%Host}" value="HOST" inline="true" checked="${instance.networking == 'HOST'}" />
+        <f:radioBlock id="networking.BRIDGE" name="networking" title="${%Bridge}" value="BRIDGE" inline="true" checked="${instance.networking == null || instance.networking == 'BRIDGE'}">
+            <f:entry title="${%Port Mappings}">
+                <f:repeatableProperty add="${%Add Port Mapping}" field="portMappings" noAddButton="false" minimum="0">
+                    <f:entry>
+                        <div align="right" class="repeatable-delete show-if-only" style="margin-left: 1em;">
+                            <f:repeatableDeleteButton value="${%Delete Port Mapping}" /><br/>
+                        </div>
+                    </f:entry>
+                </f:repeatableProperty>
+            </f:entry>
+        </f:radioBlock>
+        <f:radioBlock id="networking.USER" name="networking" title="${%User}" value="USER" inline="true" checked="${instance.networking == null || instance.networking == 'USER'}" />
+    </f:entry>
+
+    <f:entry title="${%Network Info}">
+        <f:repeatableProperty field="networkInfos" add="${%Add Network Info}" minimum="0">
+            <f:entry>
+                <div align="right" class="repeatable-delete show-if-only" style="margin-left: 1em;">
+                    <f:repeatableDeleteButton value="${%Delete Network Info}" /><br/>
+                </div>
+            </f:entry>
+        </f:repeatableProperty>
+        </f:entry>
+
+    <f:entry title="${%Volumes}">
+        <f:repeatableProperty field="volumes" add="${%Add Volume}" minimum="0">
+            <f:entry>
+                <div align="right" class="repeatable-delete show-if-only" style="margin-left: 1em;">
+                    <f:repeatableDeleteButton value="${%Delete Volume}" /><br/>
+                </div>
+            </f:entry>
+        </f:repeatableProperty>
+    </f:entry>
+
+    <f:entry title="${%Parameters}">
+        <f:repeatableProperty field="parameters" add="${%Add Parameter}" minimum="0">
+            <f:entry>
+                <div align="right" class="repeatable-delete show-if-only" style="margin-left: 1em;">
+                    <f:repeatableDeleteButton value="${%Delete Parameter}" /><br/>
+                </div>
+            </f:entry>
+        </f:repeatableProperty>
+    </f:entry>
+</j:jelly>
