+++ src/main/resources/org/jenkinsci/plugins/mesos/MesosAgentSpecTemplate/NetworkInfo/config.jelly
@@ -1,0 +1,8 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
+
+    <f:entry title="${%Name}" field="networkName">
+        <f:textbox clazz="optional"/>
+    </f:entry>
+
+</j:jelly>
