+++ src/main/resources/org/jenkinsci/plugins/mesos/MesosAgentSpecTemplate/Parameter/config.jelly
@@ -1,0 +1,10 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
+    <f:entry title="${%Name}" field="key">
+        <f:textbox field="key" clazz="required"/>
+    </f:entry>
+
+    <f:entry title="${%Value}">
+        <f:textbox field="value" clazz="required"/>
+    </f:entry>
+</j:jelly>
