+++ src/main/resources/org/jenkinsci/plugins/mesos/MesosAgentSpecTemplate/Volume/config.jelly
@@ -1,0 +1,14 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
+    <f:entry title="${%Container Path}" field="containerPath">
+        <f:textbox clazz="required"/>
+    </f:entry>
+
+    <f:entry title="${%Host Path}" field="hostPath">
+        <f:textbox/>
+    </f:entry>
+
+    <f:entry title="${%Read-Only}" field="readOnly">
+        <f:checkbox/>
+    </f:entry>
+</j:jelly>
