+++ src/test/java/org/jenkinsci/plugins/mesos/fixture/AgentSpecMother.java
@@ -3,7 +3,7 @@ package org.jenkinsci.plugins.mesos.fixture;
 import hudson.model.Node.Mode;
 import java.util.Collections;
 import org.jenkinsci.plugins.mesos.MesosAgentSpecTemplate;
-import org.jenkinsci.plugins.mesos.MesosSlaveInfo.ContainerInfo;
+import org.jenkinsci.plugins.mesos.MesosAgentSpecTemplate.ContainerInfo;
 
 /**
  * A Mother object for {@link org.jenkinsci.plugins.mesos.MesosAgentSpecTemplate}.
